Create Question1.py
A client of Expertrons wants to store the “mobile number” of all the users who are using the WhatsApp chatbot and they want us to check if the number entered by the user is “VALID” or “NOT VALID”. A mobile number is valid if: 1. It contains 10 digits. 2. Starting with 7 or 8 or 9 3. Does not contain any other character other than integers.