Create Task1.py
You have to use the News API ( https://newsapi.org ) to fetch the news on the basis of the following: 
1. Query - Suppose a user wants to search about “bitcoin” then it will show top 10 news regarding bitcoin. 
2. Source - Suppose a user wants to search a source like “techcrunch” then it will show top 10 news from techcrunch. 
The python program should ask the user to choose from option 1 and 2. If the user selects option 1 then it will show the result on the basis of “query” else “source”. 
Bonus Point (optional) Show articles of a particular “query” from a particular date.